Refactor profile page module
